Fixes several path detection issues
On windows, the column number does not seem to get reported by the backtrace crate, therefore we simply use the default value of 0 rather than requiring it in the condition.

Additionally, remove the strange "\\?\" from the beginning of canonicalised Windows paths and account for more conditions where we do not want to add an entry into the stacktrace.